+++ src/convert.c
@@ -3,35 +3,35 @@
 /*                                                        :::      ::::::::   */
 /*   convert.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: skasmi <skasmi@student.42.fr>              +#+  +:+       +#+        */
+/*   By: aelyakou <aelyakou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/27 21:20:35 by aelyakou          #+#    #+#             */
-/*   Updated: 2023/02/08 01:15:54 by skasmi           ###   ########.fr       */
+/*   Updated: 2023/02/08 19:06:36 by aelyakou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/cube3d.h"
 
-float	rad_to_deg(float x)
+double	rad_to_deg(double x)
 {
-	float	res;
+	double	res;
 
 	res = x * 180 / PI;
 	return (res);
 }
 
-float	deg_to_rad(float x)
+double	deg_to_rad(double x)
 {
-	float	res;
+	double	res;
 
 	res = x * PI / 180;
 	return (res);
 }
 
-float	normalize_ray(float rl, float ra, t_data *data)
+double	normalize_ray(double rl, double ra, t_data *data)
 {
-	float	res;
-	float	rpa;
+	double	res;
+	double	rpa;
 
 	if (ra < data->ply->pa)
 		rpa = data->ply->pa - ra;
